separate ui from api, create debug_server.py for debug use only
