+++ templates/layout/basic.jade
@@ -1,7 +1,7 @@
 html
   head
     meta(name="viewport",content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0")
-    title
+    title My Weekly Status
     script(src="{{ url_for('static', filename='lib/jquery/jquery.min.js') }}")
     script(src="{{ url_for('static', filename='js/app.js') }}")
     script(src="{{ url_for('static', filename='lib/masonry/masonry.pkgd.min.js') }}")
